consolidate date popover and date indicator into one component with the indicator as a trigger and reorganize utility functions from cardshared
