+++ frontends/main/src/app-pages/DashboardPage/CoursewareDisplay/hooks/useProgramDashboardData.ts
@@ -11,9 +11,9 @@ import {
   groupProgramEnrollmentsByProgramId,
   groupModuleCoursesByProgramId,
   buildRequirementSections,
+  getCertificateLink,
 } from "../model/dashboardViewModel"
 import type { RequirementSection } from "../model/dashboardViewModel"
-import { getCertificateLink } from "../CardShared"
 
 export type ProgramDashboardData = {
   sections: RequirementSection[]
