+++ frontends/main/src/app-pages/DashboardPage/CoursewareDisplay/model/dashboardViewModel.ts
@@ -90,6 +90,40 @@ export const EnrollmentStatus = {
 export type EnrollmentStatus =
   (typeof EnrollmentStatus)[keyof typeof EnrollmentStatus]
 
+/**
+ * Rewrites a raw mitxonline certificate link (`/certificate/{uuid}/`) to MIT
+ * Learn's own certificate route (`/certificate/{certificateType}/{uuid}/`).
+ */
+export const getCertificateLink = (
+  link: string | null | undefined,
+  certificateType: "course" | "program",
+): string | null => {
+  if (!link) return null
+  const pattern = /\/certificate\/([^/]+)\/?$/
+  return link.replace(pattern, `/certificate/${certificateType}/$1/`)
+}
+
+export const getDashboardEnrollmentStatus = (
+  resource: DashboardResource,
+): EnrollmentStatus => {
+  const hasValidCertificate =
+    resource.type !== DashboardType.Course && !!resource.data.certificate?.uuid
+
+  if (resource.type === DashboardType.Course) {
+    return EnrollmentStatus.NotEnrolled
+  }
+
+  if (resource.type === DashboardType.CourseRunEnrollment) {
+    return hasValidCertificate
+      ? EnrollmentStatus.Completed
+      : getEnrollmentStatus(resource.data)
+  }
+
+  return hasValidCertificate
+    ? EnrollmentStatus.Completed
+    : EnrollmentStatus.Enrolled
+}
+
 type KeyOpts = {
   resourceType: ResourceType
   id: number
